+++ CHANGELOG.md
@@ -48,6 +48,7 @@ All notable changes to this project will be documented in this file.
 -   Asset manager contextmenu not rendering correctly when scrolling down
 -   Grid layers of al lower floors being visible
 -   DM being able to invite themselves to the room as a player
+-   Removing a file in the asset manager now deletes the file on the server
 
 ## [0.19.3] - 2020-04-01